requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
redis==5.0.1
kafka-python==2.0.2
aiokafka==0.7.2
redis[hiredis]>=4.2.0
python-dotenv==1.0.1
httpx==0.27.0
orjson==3.9.12
loguru==0.7.2
prometheus-client==0.19.0
locust
pytest==7.4.4
pytest-asyncio==0.23.5
asyncpg==0.29.0
sqlalchemy==2.0.27
alembic==1.13.1
jinja2==3.1.2 